+++ packages/block-editor-tools/src/components/post-picker/search-modal.tsx
@@ -1,5 +1,4 @@
-import { useState } from '@wordpress/element';
-import { JSX } from 'react';
+import { useState, JSX } from 'react';
 import {
   Button,
   Modal,
@@ -39,6 +38,7 @@ const SearchModal = ({
 
   return (
     <Modal
+      className="alley-scripts-post-picker__modal"
       isDismissible
       title={modalTitle}
       onRequestClose={closeModal}