+++ packages/block-editor-tools/src/hooks/use-post-by-id/index.js
@@ -1,4 +1,4 @@
-import { useState, useEffect } from '@wordpress/element';
+import { useState, useEffect } from 'react';
 import apiFetch from '@wordpress/api-fetch';
 import { addQueryArgs } from '@wordpress/url';
 import usePost from '../use-post';
@@ -9,13 +9,16 @@ import usePost from '../use-post';
  *
  * @param {int}      postId   The ID for the post to return.
  * @param {function} getPostType Optional custom function that returns a post type string.
+ * @param {object}   options Optional object containing parameters to pass to getEntityRecord.
  * @returns {object} An object containing a hasResolved property
  *                   and the returned post object.
  */
-const usePostById = (postId, getPostType = null) => {
+const usePostById = (postId, getPostType = null, options = { context: 'view' }) => {
   const [postTypeCache, setPostTypeCache] = useState({});
+  const [working, setWorking] = useState(true);
 
   useEffect(() => {
+    setWorking(true);
     if (postId && !postTypeCache[postId]) {
       (async () => {
         if (getPostType) {
@@ -29,15 +32,22 @@ const usePostById = (postId, getPostType = null) => {
         } else {
           const path = addQueryArgs('/wp/v2/search', { include: postId });
           const newPost = await apiFetch({ path });
-          // TODO: What if the lookup fails? Should handle error state in some way.
-          setPostTypeCache((prev) => ({ ...prev, [postId]: newPost[0].subtype }));
+          if (newPost.length > 0) {
+            setPostTypeCache((prev) => ({ ...prev, [postId]: newPost[0]?.subtype }));
+          }
         }
+        setWorking(false);
       })();
+    } else if (postId && postTypeCache[postId]) {
+      setWorking(false);
     }
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [postId]);
-
-  return usePost(postId, postTypeCache[postId] ?? '');
+  const post = usePost(postId, postTypeCache[postId] ?? '', options);
+  // If we're still fetching the post type, working is true - return null.
+  // If we've got the post type, but no post yet (usePost is still working), return null.
+  // If we've got the post, return it - or return undefined as returned by usePost.
+  return working || (!post && postTypeCache[postId]) ? null : post;
 };
 
 export default usePostById;