+++ packages/block-editor-tools/webpack.config.js
@@ -2,6 +2,60 @@ const path = require('path');
 const DependencyExtractionWebpackPlugin = require('@wordpress/dependency-extraction-webpack-plugin');
 const autoprefixer = require('autoprefixer');
 
+const fs = require('fs');
+
+/**
+ * Create entry points from a directory path while maintaining the directory structure.
+ *
+ * The function will create entries from nested directories if there is an index file.
+ * The entry name will be the directory name.
+ *
+ * @param {string} directoryPath - The path to the directory to search.
+ * @returns An object of entries.
+ */
+function createEntriesPerDirPath(directoryPath) {
+  // Check if the directory exists before continuing.
+  const directoryExists = fs.existsSync(directoryPath);
+
+  if (directoryExists) {
+    // Read the directory and reduce its content to an entries object.
+    return fs.readdirSync(directoryPath)
+      .reduce((acc, dirItemName) => {
+        // Define the full path of the current directory item.
+        const fullPath = path.join(directoryPath, dirItemName);
+
+        // Define the base directory.
+        const baseDir = '/src/';
+        // Initialize an object to hold nested entries.
+        let nestedEntries = {};
+
+        if (fs.statSync(fullPath).isDirectory()) {
+          // Ensure that the directory has an index file.
+          const indexExists = ['js', 'jsx', 'ts', 'tsx']
+            .some((ext) => fs.existsSync(path.join(fullPath, `index.${ext}`)));
+
+          if (indexExists) {
+            // Get the relative path from the base directory `src/`.
+            const relativePath = fullPath.substring(fullPath.indexOf(baseDir) + baseDir.length);
+
+            // Add the current directory item to the entries object.
+            acc[dirItemName] = {
+              import: fullPath,
+              filename: `${relativePath}/index.js`,
+            };
+          }
+
+          // If current item is a directory, call the function recursively.
+          nestedEntries = createEntriesPerDirPath(fullPath);
+        }
+
+        // Return the entries object with the nested entries added
+        return { ...acc, ...nestedEntries };
+      }, {});
+  }
+  return {};
+}
+
 module.exports = (env, { mode }) => ({
   /*
    * See https://webpack.js.org/configuration/devtool/ for an explanation of how
@@ -14,7 +68,10 @@ module.exports = (env, { mode }) => ({
   devtool: mode === 'production' ? 'source-map' : 'eval-source-map',
 
   entry: {
-    index: './src',
+    // The main entry point.
+    ...{ index: './src' },
+    // Process the directories, in the 'src' directory, into their own entry points.
+    ...createEntriesPerDirPath(path.join(process.cwd(), 'src')),
   },
 
   // Configure loaders based on extension.
@@ -25,6 +82,11 @@ module.exports = (env, { mode }) => ({
         test: /\.(j|t)sx?$/,
         use: {
           loader: 'ts-loader',
+          options: {
+            compilerOptions: {
+              transpileOnly: false,
+            },
+          },
         },
       },
       {
@@ -61,7 +123,10 @@ module.exports = (env, { mode }) => ({
   // Configure plugins.
   plugins: [
     // This maps references to @wordpress/{package-name} to the wp object.
-    new DependencyExtractionWebpackPlugin(),
+    new DependencyExtractionWebpackPlugin({
+      outputFormat: 'json',
+      combineAssets: true,
+    }),
   ],
 
   resolve: {