+++ packages/create-block/CHANGELOG.md
@@ -1,11 +1,25 @@
 # Changelog
 
+## 0.2.0
+
+### Minor Changes
+
+- 71a8c3c: Allow users to opt-in to scaffolding a view script
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.1.0-next - 2023-10-02
+## 0.1.2 - 2023-12-15
+
+- Bug fix error where an import statement is used outside of a module. Ensures built files are compatible with NodeJS.
+
+## 0.1.1 - 2023-11-20
+
+- Remove `"module": "node16"` from `tsconfig.json` to inherit the default value of `"module": "esnext"`.
+
+## 0.1.0 - 2023-10-11
 
 - Update script name to `alley-create-block`.
 - Add `@wordpress/create-block` as peer dependency.