+++ packages/create-release/CHANGELOG.md
@@ -1,15 +1,28 @@
 # Changelog
 
+## 0.1.5
+
+### Patch Changes
+
+- 7008656: - Ensure that the version can be changed under a variety of different of different formats.
+  - Changes the default setting of the `--composer` and `--npm` flags to modify
+    the plugin's `composer.json`/`package.json` files if the current version of
+    the plugin is already set in the respective file.
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
 ## 0.1.3
 
 ### Patch Changes
 
 - 682b0a6: Only modify the plugin's header by default. Don't modify the `composer.json`/`package.json` version unless `--composer`/`--npm` is passed.
 
-All notable changes to this project will be documented in this file.
+## 0.1.4
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+- Consolidated shared devDependencies into the root package.json file.
 
 ## 0.1.2